Updated map views and controllers to contain markers from DB + correct infowindows

+++ application/database/migrations/2018_06_18_072927_create_poi_table.php
@@ -17,8 +17,8 @@ class CreatePoiTable extends Migration
             $table->increments('id');
             $table->string('url_id')->unique();
             $table->string('title');
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->float('latitude', 8, 6);
+            $table->float('longitude',8, 6);
             $table->string('hint')->nullable();
             $table->boolean('visibility')->default(true);
 			$table->boolean('active')->default(0);
